+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+Added
+^^^^^
+* Warn for upcoming removal of implicit targets
+
 `4.3.0`_ - 2022-11-02
 ---------------------